--- a/src/PlanetaryEphemeris.jl
+++ b/src/PlanetaryEphemeris.jl
@@ -2,7 +2,7 @@ module PlanetaryEphemeris
 
 # __precompile__(false)
 
-export PE, au, yr, sundofs, earthdofs, c_au_per_day, μ, NBP_pN_A_J23E_J23M_J2S!, NBP_pN_A_J23E_J23M_J2S_threads!, DE430!,
+export PE, au, yr, sundofs, earthdofs, c_au_per_day, μ, NBP_pN_A_J23E_J23M_J2S_threads!, DE430!,
        semimajoraxis, eccentricity, inclination, longascnode, argperi, longperi, trueanomaly, ecanomaly, meananomaly,
        timeperipass, lrlvec, eccentricanomaly, meanan2truean, meanmotion, time2truean, su, ea, mo, au, yr, daysec, clightkms,
        c_au_per_day, c_au_per_sec, c_cm_per_sec, J2000, R_sun, α_p_sun, δ_p_sun, au, UJ_interaction, de430_343ast_ids, Rx, Ry,