In order to solve https://github.com/PerezHz/PlanetaryEphemeris.jl/issues/20, this PR deprecates `taylorinteg_threads` in favor of `taylorinteg`. As a reference, a 100 steps, 4 threads integration gives the followinig benchmarks:
- main:
```julia
-----------------
Integrator warmup
-----------------
 74.505484 seconds (30.99 M allocations: 6.915 GiB, 4.37% gc time, 299.02% compilation time)
----------------
Full integration
----------------
720.649193 seconds (677.46 M allocations: 169.086 GiB, 3.20% gc time)
```
- this PR:
```julia
-----------------
Integrator warmup
-----------------
 74.704210 seconds (30.97 M allocations: 6.914 GiB, 4.33% gc time, 300.24% compilation time)
----------------
Full integration
----------------
699.725729 seconds (677.46 M allocations: 169.083 GiB, 1.97% gc time)
```
Also, results do not change between both branches.