@PerezHz 2023-07-11
Many thanks for this, @LuEdRaMo! I think this will help us avoid maintaining duplicate code, while reusing `taylorinteg`. Thank you also for posting the benchmarks, those are indeed helpful. It makes me wonder, have you tried doing runs, of say, about 10 years, in many threads, say 40 threads? I think that could help us have a better idea of the difference in performance between current `main` and this PR.