@LuEdRaMo 2023-07-12
A 10 years, 40 threads integration (in Julia v1.9.2) gives the following benchmarks:
- main:
```julia
-----------------
Integrator warmup
-----------------
102.183042 seconds (31.20 M allocations: 6.942 GiB, 2.73% gc time, 3707.83% compilation time)
----------------
Full integration
----------------
2687.964587 seconds (10.67 G allocations: 2.631 TiB, 17.19% gc time)
```
- this PR:
```julia
-----------------
Integrator warmup
-----------------
100.689653 seconds (31.19 M allocations: 6.940 GiB, 1.87% gc time, 3751.48% compilation time)
----------------
Full integration
----------------
2561.300486 seconds (10.67 G allocations: 2.631 TiB, 17.47% gc time)
```
Also, results do not change between both branches.