@PerezHz 2023-07-12
Many thanks for checking this! From this I reckon performance is pretty much the same actually... These runs include all the 343 asteroids, right?