@LuEdRaMo 2023-07-12
> These runs include all the 343 asteroids, right?

Yes, the benchmarks correspond to a full Solar System integration.