@PerezHz 2023-07-13
I'm thinking that this can be an opportunity for us to get rid of `NBP_pN_A_J23E_J23M_J2S!`, since in practice we only use the threaded version in `DE430!`. Can you try removing this function and the associated `jetcoeffs!` method?